Keep SDK test config free of Node HTTP fixtures
